requirements.txt
# Streaming
kafka-python==2.0.2
confluent-kafka==2.3.0
fastavro==1.9.4
 
# Feature Engineering
numpy==1.26.4
pandas==2.2.2
scipy==1.13.0
scikit-learn==1.4.2
feast==0.40.1
redis==5.0.1
 
# Machine Learning
lightgbm==4.3.0
torch==2.2.2
shap==0.45.0
imbalanced-learn==0.12.2
mlflow==2.12.2
 
# API
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
python-multipart==0.0.9
 
# AWS
boto3==1.34.99
aioboto3==13.0.0
 
# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
pyarrow==15.0.2
 
# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.7
 
# Monitoring
evidently==0.4.26
prometheus-client==0.20.0
 
# Utilities
python-dotenv==1.0.1
structlog==24.1.0
tenacity==8.3.0
cachetools==5.3.3
 
# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
pytest-mock==3.14.0
faker==24.14.0
